+++ spec/features/users/admin_creation_spec.rb
@@ -12,6 +12,12 @@ RSpec.feature 'Admin creation' do
     submit_profile_data(first_name: 'John', last_name: 'Smith', role: 'admin')
     expect(page).to have_content('User John Smith created.')
   end
+  it 'do not show duplicate flash messages' do
+    visit build_users_path
+    submit_username('foo@example.com')
+    click_on 'Cancel'
+    expect(page).not_to have_content('Initialized user successfully')
+  end
 
   context 'with IDR' do
     before do